requirements.txt
python-miio==0.5.6
prometheus-client==0.11.0
